+++ isitfit/cost/base_iterator.py
@@ -85,6 +85,9 @@ class BaseIterator:
     # handler of local tqdm
     self.tqdmman = tqdmman
 
+    # flag to display "Will skip ... out of ... regions ..." only once
+    self.displayed_willskip = False
+
 
   def get_regionInclude(self):
     """
@@ -157,10 +160,13 @@ class BaseIterator:
       redshift_regions_sub = [self.filter_region]
 
     # Before iterating, display a message that skipping some regions due to load from cache
-    if self.regionInclude_ready:
-      if len(redshift_regions_sub)!=len(self.region_include):
-        msg1 = "%s: Will skip %i out of %i regions to which the user/role has no access. To re-check, delete the local cache file %s"%(self.service_description, len(redshift_regions_sub)-len(self.region_include), len(redshift_regions_sub), self.simpleCacheMan.filename)
-        logger.info(colored(msg1, "yellow"))
+    # The following conditions = region_include was loaded from cache
+    if self.regionInclude_ready and len(redshift_regions_sub)!=len(self.region_include) and not self.displayed_willskip:
+        msg1 = "%s: Will skip %i out of %i regions to which the user/role has no access. To re-check, delete the local cache file %s"
+        msg1 = msg1%(self.service_description, len(redshift_regions_sub)-len(self.region_include), len(redshift_regions_sub), self.simpleCacheMan.filename)
+        import click
+        click.echo(colored(msg1, "yellow"))
+        self.displayed_willskip = True
 
     # iterate
     region_iterator = redshift_regions_sub